Rename Spotify module to Playerctl
Provides a Spotify module, that is simply a wrapper of Playerctl module,
for compatibility purposes.
